+++ CHANGELOG.md
@@ -12,11 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Pandas objects not pretty printing https://github.com/willmcgugan/rich/issues/1305
 - Fixed https://github.com/willmcgugan/rich/issues/1256
 - Fixed typing with rich.repr.auto decorator
+- Fixed repr error formatting https://github.com/willmcgugan/rich/issues/1326
 
 ### Added
 
 - Added new_line_start argument to Console.print
 - Added Segment.divide method
+- Added Segment.split_cells method
 - Added segment.SegmentLines class
 
 ## [10.4.0] - 2021-06-18